backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pdfplumber==0.11.5
docx2txt==0.8
scikit-learn==1.6.1
spacy==3.8.4
PyMySQL==1.1.1
pytest==8.3.5
